change filed

+++ server/models/plotter.py
@@ -46,8 +46,8 @@ class Plotter(models.Model):
     cpu_model = models.CharField('CPU', max_length=100, default='', blank=True)
     cpu_cnt = models.IntegerField('CPU Count', default=0)
 
-    memory_total = models.IntegerField('Memory Total', default=0)
-    memory_used = models.IntegerField('Memory Used', default=0)
+    memory_total = models.BigIntegerField('Memory Total', default=0)
+    memory_used = models.BigIntegerField('Memory Used', default=0)
